refactor: split test_flux_analysis.py into separate modules as per implementation
